+++ src/scriptlets/set-cookie-reload.js
@@ -42,6 +42,7 @@ import {
  *         - `enable` / `enabled`
  *         - `disable` / `disabled`
  *         - `necessary` / `required`
+ *         - `hide` / `hidden`
  * - `path` — optional, cookie path, defaults to `/`; possible values:
  *     - `/` — root path
  *     - `none` — to set no path at all